minor review comments
