+++ instrumentation-api/src/main/java/io/opentelemetry/instrumentation/api/instrumenter/SpanSuppressionStrategy.java
@@ -14,149 +14,37 @@ import java.util.List;
 
 abstract class SpanSuppressionStrategy {
   private static final SpanSuppressionStrategy SERVER_STRATEGY =
-      new SuppressIfSameSpanKey(singletonList(SpanKey.SERVER));
+      new SuppressIfSameSpanKeyStrategy(singletonList(SpanKey.SERVER));
   private static final SpanSuppressionStrategy CONSUMER_STRATEGY =
-      new NeverSuppressAndStore(singletonList(SpanKey.CONSUMER));
+      new NeverSuppressAndStoreStrategy(singletonList(SpanKey.CONSUMER));
   private static final SpanSuppressionStrategy ALL_CLIENTS_STRATEGY =
-      new SuppressIfSameSpanKey(singletonList(SpanKey.ALL_CLIENTS));
+      new SuppressIfSameSpanKeyStrategy(singletonList(SpanKey.ALL_CLIENTS));
   private static final SpanSuppressionStrategy ALL_PRODUCERS_STRATEGY =
-      new SuppressIfSameSpanKey(singletonList(SpanKey.ALL_PRODUCERS));
+      new SuppressIfSameSpanKeyStrategy(singletonList(SpanKey.ALL_PRODUCERS));
 
   public static final SpanSuppressionStrategy SUPPRESS_ALL_NESTED_OUTGOING_STRATEGY =
-      new CompositeStrategy(
+      new CompositeSuppressionStrategy(
           ALL_CLIENTS_STRATEGY, ALL_PRODUCERS_STRATEGY, SERVER_STRATEGY, CONSUMER_STRATEGY);
 
   private static final SpanSuppressionStrategy NO_CLIENT_SUPPRESSION_STRATEGY =
-      new CompositeStrategy(
-          NeverSuppress.INSTANCE, NeverSuppress.INSTANCE, SERVER_STRATEGY, CONSUMER_STRATEGY);
+      new CompositeSuppressionStrategy(
+          NeverSuppressOrStoreStrategy.INSTANCE,
+          NeverSuppressOrStoreStrategy.INSTANCE,
+          SERVER_STRATEGY,
+          CONSUMER_STRATEGY);
 
   static SpanSuppressionStrategy from(List<SpanKey> clientSpanKeys) {
     if (clientSpanKeys.isEmpty()) {
       return NO_CLIENT_SUPPRESSION_STRATEGY;
     }
 
-    SpanSuppressionStrategy clientOrProducerStrategy = new SuppressIfSameSpanKey(clientSpanKeys);
-    return new CompositeStrategy(
+    SpanSuppressionStrategy clientOrProducerStrategy =
+        new SuppressIfSameSpanKeyStrategy(clientSpanKeys);
+    return new CompositeSuppressionStrategy(
         clientOrProducerStrategy, clientOrProducerStrategy, SERVER_STRATEGY, CONSUMER_STRATEGY);
   }
 
   abstract Context storeInContext(Context context, SpanKind spanKind, Span span);
 
   abstract boolean shouldSuppress(Context parentContext, SpanKind spanKind);
-
-  static final class SuppressIfSameSpanKey extends SpanSuppressionStrategy {
-
-    private final List<SpanKey> outgoingSpanKeys;
-
-    SuppressIfSameSpanKey(List<SpanKey> outgoingSpanKeys) {
-      this.outgoingSpanKeys = outgoingSpanKeys;
-    }
-
-    @Override
-    Context storeInContext(Context context, SpanKind spanKind, Span span) {
-      for (SpanKey outgoingSpanKey : outgoingSpanKeys) {
-        context = outgoingSpanKey.with(context, span);
-      }
-      return context;
-    }
-
-    @Override
-    boolean shouldSuppress(Context parentContext, SpanKind spanKind) {
-      for (SpanKey outgoingSpanKey : outgoingSpanKeys) {
-        if (outgoingSpanKey.fromContextOrNull(parentContext) == null) {
-          return false;
-        }
-      }
-      return true;
-    }
-  }
-
-  static final class NeverSuppress extends SpanSuppressionStrategy {
-
-    private static final SpanSuppressionStrategy INSTANCE = new NeverSuppress();
-
-    @Override
-    Context storeInContext(Context context, SpanKind spanKind, Span span) {
-      return context;
-    }
-
-    @Override
-    boolean shouldSuppress(Context parentContext, SpanKind spanKind) {
-      return false;
-    }
-  }
-
-  static final class NeverSuppressAndStore extends SpanSuppressionStrategy {
-
-    private final List<SpanKey> outgoingSpanKeys;
-
-    NeverSuppressAndStore(List<SpanKey> outgoingSpanKeys) {
-      this.outgoingSpanKeys = outgoingSpanKeys;
-    }
-
-    @Override
-    Context storeInContext(Context context, SpanKind spanKind, Span span) {
-      for (SpanKey outgoingSpanKey : outgoingSpanKeys) {
-        context = outgoingSpanKey.with(context, span);
-      }
-      return context;
-    }
-
-    @Override
-    boolean shouldSuppress(Context parentContext, SpanKind spanKind) {
-      return false;
-    }
-  }
-
-  static final class CompositeStrategy extends SpanSuppressionStrategy {
-    private final SpanSuppressionStrategy clientStrategy;
-    private final SpanSuppressionStrategy producerStrategy;
-    private final SpanSuppressionStrategy serverStrategy;
-    private final SpanSuppressionStrategy consumerStrategy;
-
-    CompositeStrategy(
-        SpanSuppressionStrategy client,
-        SpanSuppressionStrategy producer,
-        SpanSuppressionStrategy server,
-        SpanSuppressionStrategy consumer) {
-      this.clientStrategy = client;
-      this.producerStrategy = producer;
-      this.serverStrategy = server;
-      this.consumerStrategy = consumer;
-    }
-
-    @Override
-    Context storeInContext(Context context, SpanKind spanKind, Span span) {
-      switch (spanKind) {
-        case CLIENT:
-          return clientStrategy.storeInContext(context, spanKind, span);
-        case PRODUCER:
-          return producerStrategy.storeInContext(context, spanKind, span);
-        case SERVER:
-          return serverStrategy.storeInContext(context, spanKind, span);
-        case CONSUMER:
-          return consumerStrategy.storeInContext(context, spanKind, span);
-        case INTERNAL:
-          return context;
-      }
-      return context;
-    }
-
-    @Override
-    boolean shouldSuppress(Context parentContext, SpanKind spanKind) {
-      switch (spanKind) {
-        case CLIENT:
-          return clientStrategy.shouldSuppress(parentContext, spanKind);
-        case PRODUCER:
-          return producerStrategy.shouldSuppress(parentContext, spanKind);
-        case SERVER:
-          return serverStrategy.shouldSuppress(parentContext, spanKind);
-        case CONSUMER:
-          return consumerStrategy.shouldSuppress(parentContext, spanKind);
-        case INTERNAL:
-          return false;
-      }
-      return false;
-    }
-  }
 }
